requirements.txt
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
constantly==15.1.0
cryptography==36.0.1
cssselect==1.1.0
Django==4.0
django-ckeditor==6.2.0
django-extensions==3.1.5
django-js-asset==2.0.0
h11==0.12.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
itemadapter==0.4.0
itemloaders==1.0.4
jmespath==0.10.0
lxml==4.7.1
outcome==1.1.0
parsel==1.6.0
Pillow==8.4.0
priority==1.3.0
Protego==0.1.16
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.5
pymongo==4.0.1
pyOpenSSL==21.0.0
python-dotenv==0.20.0
queuelib==1.6.2
requests==2.26.0
Scrapy==2.5.1
scrapy-selenium==0.0.7
selenium==4.1.0
service-identity==21.1.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
Twisted==21.7.0
typing_extensions==4.0.1
unicode-slugify==0.1.5
Unidecode==1.3.2
urllib3==1.26.7
w3lib==1.22.0
wsproto==1.0.0
zope.interface==5.4.0
